Users total profit on the account is displayed!

+++ app/src/main/java/com/example/csci3130project/Item.java
@@ -66,7 +66,7 @@ public class Item {
         return this.category;
     }
 
-    public double getItemValue() {
+    public int getItemValue() {
         return this.itemValue;
     }
